+++ dart/dynamics/ArrowShape.cpp
@@ -127,8 +127,7 @@ static void constructArrowTip(
 {
   std::size_t resolution = (mesh->mNumVertices - 1) / 2;
   for (std::size_t i = 0; i < resolution; ++i) {
-    double theta
-        = (double)(i) / (double)(resolution)*2 * math::constantsd::pi();
+    double theta = (double)(i) / (double)(resolution)*2 * math::pi;
 
     double R = properties.mRadius;
     double x = R * cos(theta);
@@ -156,8 +155,7 @@ static void constructArrowBody(
 {
   std::size_t resolution = mesh->mNumVertices / 2;
   for (std::size_t i = 0; i < resolution; ++i) {
-    double theta
-        = (double)(i) / (double)(resolution)*2 * math::constantsd::pi();
+    double theta = (double)(i) / (double)(resolution)*2 * math::pi;
 
     double R = properties.mRadius;
     double x = R * cos(theta);
@@ -313,16 +311,14 @@ void ArrowShape::instantiate(std::size_t resolution)
   for (std::size_t i = 0; i < resolution; ++i) {
     mesh->mNormals[2 * i].Set(0.0f, 0.0f, 1.0f);
 
-    double theta
-        = (double)(i) / (double)(resolution)*2 * math::constantsd::pi();
+    double theta = (double)(i) / (double)(resolution)*2 * math::pi;
     mesh->mNormals[2 * i + 1].Set(cos(theta), sin(theta), 0.0f);
   }
   mesh->mNormals[mesh->mNumVertices - 1].Set(0.0f, 0.0f, -1.0f);
 
   mesh = scene->mMeshes[1];
   for (std::size_t i = 0; i < resolution; ++i) {
-    double theta
-        = (double)(i) / (double)(resolution)*2 * math::constantsd::pi();
+    double theta = (double)(i) / (double)(resolution)*2 * math::pi;
     mesh->mNormals[2 * i].Set(cos(theta), sin(theta), 0.0f);
     mesh->mNormals[2 * i + 1].Set(cos(theta), sin(theta), 0.0f);
   }
@@ -331,8 +327,7 @@ void ArrowShape::instantiate(std::size_t resolution)
   for (std::size_t i = 0; i < resolution; ++i) {
     mesh->mNormals[2 * i].Set(0.0f, 0.0f, -1.0f);
 
-    double theta
-        = (double)(i) / (double)(resolution)*2 * math::constantsd::pi();
+    double theta = (double)(i) / (double)(resolution)*2 * math::pi;
     mesh->mNormals[2 * i + 1].Set(cos(theta), sin(theta), 0.0f);
   }
   mesh->mNormals[mesh->mNumVertices - 1].Set(0.0f, 0.0f, 1.0f);